requirements.txt
torch>=2.3.0
transformers>=4.45.0
peft>=0.12.0
accelerate>=0.34.0
kagglehub>=0.3.0
polars>=1.0.0
datasets>=3.0.0
trl>=0.11.0
scipy>=1.12.0
sentencepiece>=0.2.0
protobuf>=4.25.0
safetensors>=0.4.0
wandb>=0.18.0
PyYAML>=6.0
requests>=2.32.0
